\name{run.jagsfile}

\alias{run.jagsfile}
\alias{run.JAGSfile}
\alias{autorun.jagsfile}
\alias{autorun.JAGSfile}

\alias{xgrid.autorun.jagsfile}
\alias{xgrid.autorun.JAGSfile}
\alias{xgrid.run.jagsfile}
\alias{xgrid.run.JAGSfile}
\alias{xgrid.submit.jagsfile}
\alias{xgrid.submit.JAGSfile}


\title{Deprecated functions}

\description{

These functions are deprecated and only remain for backwards compatibility.  
Please use the run.jags or autorun.jags functions directly, with the model argument replacing path.

}


\seealso{

\code{\link{autorun.jags}},

\code{\link{run.jags}},

\code{\link{xgrid.run.jags}},

\code{\link{read.winbugs}}

}



\author{Matthew Denwood \email{matthew.denwood@glasgow.ac.uk}}


\keyword{models}