% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runjags-package.R
\docType{package}
\name{runjags-package}
\alias{runjags}
\alias{runjags-package}
\title{runjags: Interface Utilities, Model Templates, Parallel Computing Methods and Additional Distributions for MCMC Models in JAGS}
\description{
User-friendly interface utilities for MCMC models via
    Just Another Gibbs Sampler (JAGS), facilitating the use of parallel
    (or distributed) processors for multiple chains, automated control
    of convergence and sample length diagnostics, and evaluation of the
    performance of a model using drop-k validation or against simulated
    data. Template model specifications can be generated using a standard
    lme4-style formula interface to assist users less familiar with the
    BUGS syntax.  A JAGS extension module provides additional distributions
    including the Pareto family of distributions, the DuMouchel prior and
    the half-Cauchy prior.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ku-awdc/runjags}
  \item Report bugs at \url{https://github.com/ku-awdc/runjags/issues}
}

}
\author{
\strong{Maintainer}: Matthew Denwood \email{md@sund.ku.dk}

Other contributors:
\itemize{
  \item Martyn Plummer (Copyright holder of the code in src/distributions/DPar1.*, configure.ac, R/rjags_functions.R, and original copyright holder of some modified code where indicated) [copyright holder]
}

}
\keyword{internal}
