% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_huiwalter.R
\name{template_huiwalter}
\alias{template_huiwalter}
\title{Title}
\usage{
template_huiwalter(
  testdata,
  outfile = "huiwalter_model.txt",
  covon = FALSE,
  se_priors = "dbeta(1,1)",
  sp_priors = "dbeta(1,1)"
)
}
\arguments{
\item{sp_priors}{}
}
\description{
Title
}
